+++ R/fst_hudson_k.R
@@ -11,6 +11,11 @@
 #' @param ind_keep An optional vector of individuals to keep (as booleans or indexes, used to subset an R matrix).
 #' @param loci_on_cols Determines the orientation of the genotype matrix (by default, `FALSE`, loci are along the rows).
 #' Set automatically to `TRUE` if `X` is a BEDMatrix object.
+#' @param mem_factor Proportion of available memory to use loading and processing genotypes.
+#' Ignored if `mem_lim` is not `NA`.
+#' @param mem_lim Memory limit in GB, used to break up genotype data into chunks for very large datasets.
+#' Note memory usage is somewhat underestimated and is not controlled strictly.
+#' Default in Linux and Windows is `mem_factor` times the free system memory, otherwise it is 1GB (OSX and other systems).
 #'
 #' @return A list with the following named elements, in this order:
 #' 
@@ -60,7 +65,7 @@
 #' The popkin package.
 #'
 #' @export
-fst_hudson_k <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE) {
+fst_hudson_k <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE, mem_factor = 0.7, mem_lim = NA) {
     if (missing(X))
         stop('Genotype matrix `X` is required!')
     if (missing(labs))
@@ -94,8 +99,17 @@ fst_hudson_k <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE)
     FstTs <- vector('numeric', m)
     FstBs <- vector('numeric', m)
 
+    # calculate chunk size given available memory
     # use the same formula as for WC (the largest matrices are the same size)
-    mc <- get_mem_lim_m_WC(n = n, r = r, m = m)
+    data <- popkin:::solve_m_mem_lim(
+                         n = 1.5 * n + 2 * r,
+                         m = m,
+                         mat_m_n = 1, # left simple to not confound with above `n` hack
+                         vec_m = 6,
+                         mem = mem_lim,
+                         mem_factor = mem_factor
+                     )
+    mc <- data$m_chunk
     
     # navigate chunks
     mci <- 1 # start of first chunk (needed for matrix inputs only; as opposed to function inputs)