+++ R/fst_wc.R
@@ -11,6 +11,11 @@
 #' @param ind_keep An optional vector of individuals to keep (as booleans or indexes, used to subset an R matrix).
 #' @param loci_on_cols Determines the orientation of the genotype matrix (by default, `FALSE`, loci are along the rows).
 #' Set automatically to `TRUE` if `X` is a BEDMatrix object.
+#' @param mem_factor Proportion of available memory to use loading and processing genotypes.
+#' Ignored if `mem_lim` is not `NA`.
+#' @param mem_lim Memory limit in GB, used to break up genotype data into chunks for very large datasets.
+#' Note memory usage is somewhat underestimated and is not controlled strictly.
+#' Default in Linux and Windows is `mem_factor` times the free system memory, otherwise it is 1GB (OSX and other systems).
 #'
 #' @return A list with the following named elements, in this order:
 #' 
@@ -62,7 +67,7 @@
 #' The popkin package.
 #'
 #' @export
-fst_wc <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE) {
+fst_wc <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE, mem_factor = 0.7, mem_lim = NA) {
     if (missing(X))
         stop('Genotype matrix `X` is required!')
     if (missing(labs))
@@ -103,8 +108,35 @@ fst_wc <- function(X, labs, m = NA, ind_keep = NULL, loci_on_cols = FALSE) {
     FstBs <- vector('numeric', m)
     mafs <- vector('numeric', m)
     
+    # estimating total memory usage in bytes
+    # Pi = m*8+ao
+    # k2ps = m*r*8+mo
+    # Xi[, k2is[[k]] ] = m*(n/r)*8+mo # on average, let's be conservative: m*n*8+mo
+    # (k2ps-Pi)^2 = m*r*8+mo # intermediate calculation, unnamed?
+    # s2s = m*8+ao # final length
+    # nus = m*8+ao # another vector
+    # Xi == 1 = m*n*4+mo # intermediate matrix, one of the bigger ones!
+    # hbars = m*8+ao
+    # FstTsi = m*8+ao
+    # FstBsi = m*8+ao
+
+    # the general function doesn't have an "r" input, but since "n" only enters in m*n matrices and so does "r", we can hack it all to have an effective "n" that combines both
+    # Have:
+    # - 1.5 m*n (input int matrix, plus 1 subpop extract)
+    # - 2 m*r
+    # since there's no other n's this is equivalent to
+    # n = 1.5 * n + 2 * r
+    
     # calculate chunk size given available memory
-    mc <- get_mem_lim_m_WC(n = n, r = r, m = m)
+    data <- popkin:::solve_m_mem_lim(
+                         n = 1.5 * n + 2 * r,
+                         m = m,
+                         mat_m_n = 1, # left simple to not confound with above `n` hack
+                         vec_m = 6,
+                         mem = mem_lim,
+                         mem_factor = mem_factor
+                     )
+    mc <- data$m_chunk
     
     # navigate chunks
     mci <- 1 # start of first chunk (needed for matrix inputs only; as opposed to function inputs)