+++ NEWS.md
@@ -22,3 +22,7 @@
 
 * Adapted package to use new `solve_m_mem_lim` from the `popkin` package, which generalizes some of the calculations that used to be part of the memory estimation/control code in `get_mem_lim_m_kinship_std` and `get_mem_lim_m_WC`.
 * `kinship_std` now preserves individual names when present in input genotype matrix.
+
+# 2019-08-02 - popkinsuppl 1.0.5.9000
+
+* Internal changes to match updated `solve_m_mem_lim` from the `popkin` package (>= 1.2.6.9000).